crime_analysis/evaluation/metrics.py
"""
評估指標模組
四種實驗的主要指標：
  對照實驗 I:  AUC, COMET, ROUGE-L
  消融實驗 II: F1-score, BLEU
  驗證實驗 III: 收斂速度 (Epochs), 決策效率 (Steps)
  專家驗證 IV:  MOS (Mean Opinion Score)
"""
from typing import Any, Dict, List, Optional
import logging

logger = logging.getLogger(__name__)


class MetricsCalculator:
    """統一的評估指標計算器"""

    # ── 傳統詞彙重疊指標 ─────────────────────────────────

    def compute_rouge_l(self, hypothesis: str, reference: str) -> float:
        """
        ROUGE-L：最長公共子序列（LCS）
        衡量報告的詞彙覆蓋率

        TODO: 安裝 rouge-score
        from rouge_score import rouge_scorer
        scorer = rouge_scorer.RougeScorer(['rougeL'], use_stemmer=False)
        scores = scorer.score(reference, hypothesis)
        return scores['rougeL'].fmeasure
        """
        return 0.0  # placeholder

    def compute_bleu(
        self, hypothesis: str, references: List[str], n: int = 4
    ) -> float:
        """
        BLEU-N：N-gram 精確度
        衡量報告與參考答案的字面相似度

        TODO:
        from nltk.translate.bleu_score import sentence_bleu, SmoothingFunction
        ref_tokens = [ref.split() for ref in references]
        hyp_tokens = hypothesis.split()
        return sentence_bleu(ref_tokens, hyp_tokens,
                             smoothing_function=SmoothingFunction().method1)
        """
        return 0.0  # placeholder

    # ── 語意相似度指標 ────────────────────────────────────

    def compute_bertscore(
        self, hypothesis: str, reference: str, lang: str = "zh"
    ) -> Dict[str, float]:
        """
        BERTScore：基於 BERT 的語意相似度
        比 BLEU 更能捕捉語意層面的品質

        TODO:
        from bert_score import score
        P, R, F1 = score([hypothesis], [reference], lang=lang, verbose=False)
        return {"precision": P.item(), "recall": R.item(), "f1": F1.item()}
        """
        return {"precision": 0.0, "recall": 0.0, "f1": 0.0}  # placeholder

    def compute_comet(
        self,
        source: str,      # 影像描述（來源）
        hypothesis: str,  # 系統生成的鑑定報告
        reference: str,   # 人類專家的參考報告
    ) -> float:
        """
        COMET：基於神經網路的翻譯/生成品質評估
        原為機器翻譯設計，但適用於任何生成任務

        TODO:
        from comet import download_model, load_from_checkpoint
        model_path = download_model("Unbabel/wmt22-comet-da")
        model = load_from_checkpoint(model_path)
        data = [{"src": source, "mt": hypothesis, "ref": reference}]
        scores = model.predict(data, batch_size=1)
        return scores.scores[0]
        """
        return 0.0  # placeholder

    # ── 分類指標 ──────────────────────────────────────────

    def compute_classification_metrics(
        self,
        predictions: List[str],
        ground_truths: List[str],
        categories: List[str],
    ) -> Dict[str, Any]:
        """
        計算多類別分類指標：
        - Accuracy
        - Macro F1
        - Per-class F1
        - AUC（需要機率分數）

        TODO:
        from sklearn.metrics import (
            accuracy_score, f1_score,
            classification_report, roc_auc_score
        )
        acc = accuracy_score(ground_truths, predictions)
        macro_f1 = f1_score(ground_truths, predictions, average='macro')
        report = classification_report(ground_truths, predictions,
                                       target_names=categories, output_dict=True)
        return {"accuracy": acc, "macro_f1": macro_f1, "per_class": report}
        """
        correct = sum(p == g for p, g in zip(predictions, ground_truths))
        accuracy = correct / len(predictions) if predictions else 0.0
        return {"accuracy": accuracy, "macro_f1": 0.0, "per_class": {}}

    # ── 消融實驗指標 ──────────────────────────────────────

    def compute_ablation_table(
        self,
        results_by_config: Dict[str, List[Dict]],
        ground_truths: List[str],
    ) -> Dict[str, Dict]:
        """
        Leave-One-Out 消融實驗結果彙整
        configs 例：
            "full_system" / "no_action" / "no_environment" /
            "no_time_emotion" / "no_semantic" / "no_reflector"

        TODO: 對每個 config 計算所有指標並彙整為表格
        """
        ablation_results = {}
        for config_name, preds_list in results_by_config.items():
            preds = [r.get("final_category", "Normal") for r in preds_list]
            metrics = self.compute_classification_metrics(
                preds, ground_truths, categories=[]
            )
            ablation_results[config_name] = metrics

        return ablation_results

    # ── 強化學習訓練指標 ──────────────────────────────────

    def compute_convergence_metrics(
        self, training_log: List[Dict]
    ) -> Dict[str, Any]:
        """
        驗證實驗 III：強化學習收斂速度
        - 達到 Racc > 0.8 所需的步數
        - 平均 Rcons 曲線
        - 平均對話輪次趨勢（決策效率）
        """
        if not training_log:
            return {}

        rewards = [entry.get("mean_reward", 0) for entry in training_log]
        rcons_vals = [entry.get("rcons", 0) for entry in training_log]

        # 找到首次超過 0.8 獎勵的步數
        convergence_step = next(
            (i for i, r in enumerate(rewards) if r > 0.8),
            len(rewards),
        )

        return {
            "convergence_step": convergence_step,
            "final_mean_reward": rewards[-1] if rewards else 0.0,
            "mean_rcons": sum(rcons_vals) / len(rcons_vals) if rcons_vals else 0.0,
            "total_steps": len(training_log),
        }
