crime_analysis/requirements.txt
# ── 核心模型 ──────────────────────────────────────────
transformers>=4.45.0
torch>=2.3.0
peft>=0.13.0               # LoRA 微調（Semantic Agent）
trl>=0.11.0                # GRPO / DPO 訓練框架

# ── 視覺語言模型 ──────────────────────────────────────
qwen-vl-utils              # Qwen2.5-VL 輔助工具
openai>=1.50.0             # Planner (GPT-4o) & DPO Judge

# ── 影像處理 ──────────────────────────────────────────
opencv-python>=4.10.0
mediapipe>=0.10.0          # 33 個姿勢關鍵點（Time&Emotion Agent）
deepface>=0.0.93           # 情緒分析（Time&Emotion Agent）

# ── RAG 系統 ──────────────────────────────────────────
chromadb>=0.5.0            # 向量資料庫（Dense Layer）
sentence-transformers>=3.1.0  # BGE-M3 Embedding
rank-bm25>=0.2.2           # BM25 稀疏檢索（Keyword Layer）
jieba>=0.42.1              # 中文分詞（BM25 前處理）

# ── 評估指標 ──────────────────────────────────────────
rouge-score>=0.1.2         # ROUGE-L
nltk>=3.9.1                # BLEU
bert-score>=0.3.13         # BERTScore
unbabel-comet>=2.2.2       # COMET
scikit-learn>=1.5.0        # AUC, F1, Classification Report

# ── 工具 ──────────────────────────────────────────────
numpy>=1.26.0
pandas>=2.2.0
tqdm>=4.66.0
pydantic>=2.8.0
python-dotenv>=1.0.0       # API Key 管理
